Move config-class to better location
